+++ delegate.js
@@ -1,0 +1,2 @@
+import Delegate from './dist/delegate';
+module.exports = Delegate;