+++ dquery.js
@@ -1,0 +1,2 @@
+import DQuery from './dist/dquery';
+module.exports = DQuery;