+++ vdom.js
@@ -1,0 +1,2 @@
+import VDOM from './dist/vdom';
+module.exports = VDOM;